No også tokengenerering for systembrukar
