+++ apps/etterlatte-grunnlag/src/test/kotlin/grunnlag/BehandlingGrunnlagRoutesKtTest.kt
@@ -32,6 +32,7 @@ import io.mockk.verify
 import lagGrunnlagsopplysning
 import no.nav.etterlatte.grunnlag.klienter.BehandlingKlient
 import no.nav.etterlatte.ktor.issueSaksbehandlerToken
+import no.nav.etterlatte.ktor.issueSystembrukerToken
 import no.nav.etterlatte.libs.common.behandling.SakType
 import no.nav.etterlatte.libs.common.grunnlag.Grunnlagsopplysning
 import no.nav.etterlatte.libs.common.grunnlag.NyeSaksopplysninger
@@ -85,19 +86,6 @@ internal class BehandlingGrunnlagRoutesKtTest {
     @AfterAll
     fun after() {
         server.shutdown()
-    }
-
-    private val systemBruker: String by lazy {
-        val mittsystem = UUID.randomUUID().toString()
-        server.issueToken(
-            issuerId = AZURE_ISSUER,
-            audience = CLIENT_ID,
-            claims =
-                mapOf(
-                    "sub" to mittsystem,
-                    "oid" to mittsystem,
-                ),
-        ).serialize()
     }
 
     @Test
@@ -295,7 +283,7 @@ internal class BehandlingGrunnlagRoutesKtTest {
                     contentType(ContentType.Application.Json)
                     setBody(opplysningsbehov)
                     headers {
-                        append(HttpHeaders.Authorization, "Bearer $systemBruker")
+                        append(HttpHeaders.Authorization, "Bearer ${server.issueSystembrukerToken()}")
                     }
                 }
 
@@ -328,7 +316,7 @@ internal class BehandlingGrunnlagRoutesKtTest {
                     contentType(ContentType.Application.Json)
                     setBody(request)
                     headers {
-                        append(HttpHeaders.Authorization, "Bearer $systemBruker")
+                        append(HttpHeaders.Authorization, "Bearer ${server.issueSystembrukerToken()}")
                     }
                 }
